requirements.txt
torch
torchvision
timm
numpy
pandas
scikit-learn
matplotlib
seaborn
opencv-python
PyWavelets
scikit-image
tqdm
Pillow
kaggle
torchmetrics
kagglehub
